feat(search): integrate search functionality into HomeController and remove SearchController
