+++ src/controllers/HomeController.ts
@@ -18,7 +18,7 @@ export class HomeController {
     getResourcesForUser = async (req: AuthenticatedRequest, res: Response): Promise<void> => {
         try {
             const userId = req.user.id;
-            
+
             const [decks, classes, folders] = await Promise.all([
                 this.deckRepository.findByUserId(userId),
                 this.classRepository.findByUserId(userId),
@@ -36,6 +36,29 @@ export class HomeController {
             res.status(500).json({
                 message: error || 'Internal server error',
             });
+        }
+    };
+
+    search = async (req: AuthenticatedRequest, res: Response): Promise<void> => {
+        try {
+            const keyword = req.query.keyword as string;
+            const limit = parseInt(req.query.limit as string) || 10;
+
+            if (!keyword) {
+                res.status(400).json({ message: 'Missing search keyword' });
+                return;
+            }
+
+            const decks = await this.deckRepository.findByKeyword(keyword);
+
+            res.status(200).json({
+                data: {
+                    decks,
+                },
+            });
+        } catch (error) {
+            console.error('Error during search:', error);
+            res.status(500).json({ message: 'Internal server error' });
         }
     };
 }
